fix: Player no longer gets stuck in the ground
Previously, when the player landed from a jump they would get stuck
in the ground (couldn't move left or right at all). The only way to get
unstuck was to jump again, thus repeating the cycle.

This prompted a redesign to how movement and collisons are handled.
Instead of moving the player one pixel in every direction to check for
collision, the new system moves a copy of the player's rectangle to
check that the proposed move is valid and won't cause any collisions.

Additionally, if the copy does collide with something, rather than
rejecting the move altogether the system will move the player by the max
distance possible that doesn't cause collsion.

Fixes:
- on_ground status is now accurately reported when player is stationary
or moving on the ground tiles
- Player no longer collides with the ground during left/right movement

New bugs:
- If gravity is applied when on_ground == False, the player floats in
in the air when sliding off of platforms.
- If gravity is always applied, the player's vertical speed is briefly
non-zero when on the ground (might cause problems in the future).
- Player teleports when colliding with more than one entity in the left/
up directions.