fix: player no longer teleports during collision
I found the issue! Despite the move and move_ip methods accepting
float type parameters, pygame will only ever move a sprite's rectangle
by an integer amount.

We already discovered this when implementing the initial version of the
movement code, with the fix being to round the movement distance down
to the nearest integer. But what I failed to realize was that because
the refactored collision system also involves moving a rectangle
(albeit a copy of an entity's rectangle), the same issue would pop up.
So after all that, the fix to the mysterious teleportation bug ended up
being one we already used somewhere else before.