fix: applied correct hitboxes to spikes
To achieve this I added a new hitbox attribute to the entity class. By
default the hixbox uses the same rectangle as the image, but for the
spikes we shrink the hitbox and move it to match the true size/position
of the spikes. This hitbox rect is then used in collisions.

Player and portal now pull their animation speed, gravity, etc. from the
config file instead of relying on hard-coded values in the class.