ci: added GitHub action that lints the code
Ruff is run twice in the action, the first checks for any errors that
would prevent the code from running and the second checks for style,
type hinting, etc. issues.

Also added args to the ruff pre-commit hook so that it only fails when
a syntax or other runtime error is found.